requirements.txt
paho-mqtt==1.6.1
boto3==1.34.0
python-dotenv==1.0.0